requirements.txt
flask==2.3.3
langchain-core>=0.3.42,<0.4.0
langchain>=0.1.0
langgraph>=0.0.15
chromadb==0.4.22
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
langchain-groq==0.2.5
trafilatura==1.6.1
python-whois==0.8.0
openai==1.3.5
tavily-python==0.2.6
langchain-community>=0.0.10
